pyproject.toml
[build-system]
requires = ["setuptools>=64", "setuptools-scm>=8"]
build-backend = "setuptools.build_meta"

[tool.setuptools_scm]
# can be empty if no extra settings are needed, presence enables setuptools-scm

[project.urls]
Homepage = "https://github.com/srmaschek/QUARK-plugin-myqlm"

[project]
name = "QUARK-plugin-myqlm"
dynamic = ["version"]
readme = "README.md"
requires-python = ">=3.12"
classifiers = [
    "Programming Language :: Python :: 3",
    "Operating System :: OS Independent",
]
license = "Apache-2.0"
dependencies = ["quark-framework>=0.3.7", "myqlm==1.10.2"]
